+++ Install.sh
@@ -182,6 +182,6 @@ rm -rf build
 mkdir build
 cd build
 cmake ../ ${FLAG}
-make -j`nproc`
+make -j4
 make install
 # ctest
