+++ developer_tools/clang-formater.sh
@@ -1,0 +1,3 @@
+var=$(find . -iname *.hpp -o -iname *.h -o -iname *.cpp -o -iname *.cu)
+echo $var
+clang-format -i $var -style=file
