+++ include/Tensor.old.hpp
@@ -17,7 +17,7 @@
 #include <vector>
 #include <initializer_list>
 #include <string>
-#include "Scalar.hpp"
+#include "backend/Scalar.hpp"
 
 namespace cytnx {
