+++ include/UniTensor.hpp
@@ -5,7 +5,7 @@
 #include "cytnx_error.hpp"
 #include "Device.hpp"
 #include "Tensor.hpp"
-#include "Scalar.hpp"
+#include "backend/Scalar.hpp"
 #include "utils/utils.hpp"
 #include "intrusive_ptr_base.hpp"
 #include <iostream>
