+++ include/utils/print.hpp
@@ -8,7 +8,7 @@
 #include "Symmetry.hpp"
 #include "UniTensor.hpp"
 #include "Network.hpp"
-#include "Scalar.hpp"
+#include "backend/Scalar.hpp"
 #include <iostream>
 
 namespace cytnx {
