+++ src/CMakeLists.txt
@@ -6,7 +6,7 @@
 
 target_sources_local(cytnx
   PRIVATE
-    Scalar.cpp
+    #Scalar.cpp
     Network_base.cpp
     Network.cpp
     RegularNetwork.cpp
