+++ src/algo/Hsplit.cpp
@@ -3,7 +3,7 @@
 #include "Accessor.hpp"
 #include "Generator.hpp"
 #include "backend/Storage.hpp"
-#include "Scalar.hpp"
+#include "backend/Scalar.hpp"
 namespace cytnx {
   namespace algo {
     typedef Accessor ac;
