+++ src/linalg/Ger.cpp
@@ -2,7 +2,7 @@
 #include "linalg_internal_interface.hpp"
 #include "utils/utils.hpp"
 #include "Tensor.hpp"
-#include "Scalar.hpp"
+#include "backend/Scalar.hpp"
 #include "Generator.hpp"
 
 namespace cytnx {
